requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.1
python-multipart==0.0.9
pydantic==2.6.1
pillow==10.2.0